[feat][operator] validate Ray resource metadata in webhook (#3831)
when the Ray resource is created or updated. For users who opt into installing
the webhooks, this validation makes the K8s API fail the resource creation or
update request immediately in a synchronous way that's more obvious.

We refactor the RayCluster webhook to use the same validation functions that
the RayCluster controller uses. We also add tests for RayJob and RayService webhooks.

Signed-off-by: David Xia <david@davidxia.com>